# Speed up mapping creation/activation