--- a/frontend/javascripts/oxalis/default_state.js
+++ b/frontend/javascripts/oxalis/default_state.js
@@ -81,6 +81,7 @@ const defaultState: OxalisState = {
     activeMapping: {
       mappingName: null,
       mapping: null,
+      mappingKeys: null,
       mappingColors: null,
       hideUnmappedIds: false,
       isMappingEnabled: false,