--- a/frontend/javascripts/oxalis/store.js
+++ b/frontend/javascripts/oxalis/store.js
@@ -280,6 +280,7 @@ export type TemporaryConfiguration = {
   +activeMapping: {
     +mappingName: ?string,
     +mapping: ?Mapping,
+    +mappingKeys: ?Array<number>,
     +mappingColors: ?Array<number>,
     +hideUnmappedIds: boolean,
     +isMappingEnabled: boolean,