… frontend/javascripts/test/screenshots/ROI2017_wkw_with_mapping_axons.png